Refactor components into files
